+++ packages/slick-components/src/index.js
@@ -1,2 +1,3 @@
 export * from './Arrow'
 export * from './Dots'
+export * from './SlickSlider'